+++ Goatshop/SupplierForm.cs
@@ -92,16 +92,24 @@ namespace Goatshop
         // Delete
         private void buttonDelete_Click(object sender, EventArgs e)
         {
-            if (SelectedSupp().Product.Count == 0)
+            if (listViewSupplier.SelectedItems.Count > 0)
             {
-                Settings.db.Supplier.Remove(SelectedSupp());
-                Settings.db.SaveChanges();
 
-                ShowList();
+                if (SelectedSupp().Product.Count == 0)
+                {
+                    Settings.db.Supplier.Remove(SelectedSupp());
+                    Settings.db.SaveChanges();
+
+                    ShowList();
+                }
+                else
+                {
+                    MessageBox.Show("Sorry you cant delete a Supplier with attached products. \nPlease remove or change those products first.\n\nTry again after changing", "Error!");
+                }
             }
             else
             {
-                MessageBox.Show("Sorry you cant delete a Supplier with attached products. \nPlease remove or change those products first.\n\nTry again after changing", "Error!");
+                MessageBox.Show("Nothing selected", "Info");
             }
         }
